+++ material-core/src/main/java/cn/enilu/material/dao/BaseRepository.java
@@ -1,5 +1,6 @@
 package cn.enilu.material.dao;
 
+import cn.enilu.material.bean.vo.query.SearchFilter;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.NoRepositoryBean;
@@ -7,6 +8,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import java.io.Serializable;
 import java.util.List;
+import java.util.Map;
 
 /**
  * 封装基础的dao接口
@@ -18,7 +20,8 @@ import java.util.List;
 public interface BaseRepository<T, ID extends Serializable> extends JpaRepository<T, ID>
         , PagingAndSortingRepository<T, ID>
         , JpaSpecificationExecutor<T> {
-    List<Object[]> queryBySql(String sql);
+    List<Map> queryBySql(String sql);
+    List<Map> queryBySql(String sql, List<SearchFilter> filter);
     List<T> query(String sql);
     Object getBySql(String sql);
     T get(String sql);