+++ material-core/src/main/java/cn/enilu/material/dao/BaseRepositoryImpl.java
@@ -1,11 +1,16 @@
 package cn.enilu.material.dao;
 
+import cn.enilu.material.bean.vo.query.SearchFilter;
+import org.hibernate.SQLQuery;
+import org.hibernate.transform.Transformers;
 import org.springframework.data.jpa.repository.support.JpaEntityInformation;
 import org.springframework.data.jpa.repository.support.SimpleJpaRepository;
 
 import javax.persistence.EntityManager;
+import javax.persistence.Query;
 import java.io.Serializable;
 import java.util.List;
+import java.util.Map;
 
 /**
  * 基础dao实现类
@@ -28,8 +33,21 @@ public class BaseRepositoryImpl<T, ID extends Serializable>
     }
 
     @Override
-    public List<Object[]> queryBySql(String sql) {
-        return entityManager.createNativeQuery(sql).getResultList();
+    public List<Map> queryBySql(String sql) {
+        return queryBySql(sql,null);
+    }
+
+    @Override
+    public List<Map> queryBySql(String sql, List<SearchFilter> filters) {
+        Query query = entityManager.createNativeQuery(sql);
+        if(filters!=null&&!filters.isEmpty()){
+            for(SearchFilter filter:filters){
+                query.setParameter(filter.fieldName, filter.value);
+            }
+        }
+        query.unwrap(SQLQuery.class).setResultTransformer(Transformers.ALIAS_TO_ENTITY_MAP);
+        List list = query.getResultList();
+        return list;
     }
 
     @Override