+++ material-core/src/main/java/cn/enilu/material/dao/system/MenuRepository.java
@@ -17,8 +17,6 @@ public interface MenuRepository extends BaseRepository<Menu,Long> {
     Menu findByCode(String code);
     List<Menu> findByPcodesLike(String code);
     List<Menu> findByNameLikeAndLevels(String name,Integer levels);
-    @Query(nativeQuery = true,value = "select url from t_sys_relation rel inner join t_sys_menu m on rel.menuid = m.id where m.status=1 and  rel.roleid = ?1")
-    List<String> getResUrlsByRoleId(Integer roleId);
     @Query(nativeQuery = true,value="SELECT m1.id AS id, m1.icon AS icon, ( CASE WHEN (m2.id = 0 OR m2.id IS NULL) " +
             "THEN 0 ELSE m2.id END ) AS parentId, m1. NAME AS NAME, m1.url AS url, m1.levels AS levels, m1.ismenu AS " +
             "ismenu, m1.num AS num,m1.code as code,m1.status as status FROM t_sys_menu m1 LEFT JOIN t_sys_menu m2 ON " +
@@ -41,9 +39,6 @@ public interface MenuRepository extends BaseRepository<Menu,Long> {
     @Query(nativeQuery = true,value = "SELECT m1.id AS id, ( CASE WHEN (m2.id = 0 OR m2.id IS NULL) THEN 0 ELSE m2.id END ) AS pId, m1. NAME AS NAME, ( CASE WHEN (m2.id = 0 OR m2.id IS NULL) THEN 'true' ELSE 'false' END ) AS isOpen, ( CASE WHEN (m3.ID = 0 OR m3.ID IS NULL) THEN 'false' ELSE 'true' END ) \"checked\" FROM t_sys_menu m1 LEFT JOIN t_sys_menu m2 ON m1.pcode = m2. CODE LEFT JOIN ( SELECT ID FROM t_sys_menu WHERE ID IN (?1)) m3 ON m1.id = m3.id ORDER BY m1.id ASC")
     List menuTreeListByMenuIds(List<Long> menuIds);
 
-    List<Menu> findByNameLike(String name);
-
-    List<Menu> findByLevels(Integer level);
     @Modifying
     @Transactional
     @Query(nativeQuery = true,value = "delete from t_sys_relation where menuid=?1")